Remove keystore during build
Overwriting the keystore during a new deployment was preventing
GeoServer from starting. My testing seems to indicate that GeoServer
will create this on startup if it doesn't exist, so there shouldn't be a
problem with removing it. Honestly, I'm not really even sure if we're
using this; it's hard to tell from the docs.